Move Metric Server to Jobs Dyno
