+++ jobs/index.ts
@@ -7,11 +7,14 @@ import { getLogger } from '../common/logger/logger';
 import { scheduleJobs } from './scheduler';
 import * as scheduler from './scheduler';
 import { configureGracefulShutdown } from './shutdown';
+import { jobExists, regularJobs } from './list';
+import { runJob } from './execute';
+import express from 'express';
+import { metricsRouter } from '../common/logger/metrics';
+import http from 'http';
 
 // Ensure Notification hooks are always loaded
 import './../common/notification/hooks';
-import { jobExists, regularJobs } from './list';
-import { runJob } from './execute';
 
 //SETUP: logger
 const log = getLogger();
@@ -23,6 +26,24 @@ moment.tz.setDefault('Europe/Berlin'); //set global timezone (which is then used
 
 //SETUP: Add a graceful shutdown to the scheduler used
 configureGracefulShutdown(scheduler);
+
+// Add Metrics Server to Jobs Dyno
+async function startMetricsServer() {
+    const app = express();
+    app.use('/metrics', metricsRouter);
+
+    const port = process.env.PORT || 5100;
+
+    const server = http.createServer(app);
+
+    // Start listening
+    await new Promise<void>((res) => server.listen(port, res));
+    log.info(`Server listening on port ${port}`);
+}
+
+if (process.env.METRICS_SERVER_ENABLED === 'true') {
+    startMetricsServer().catch((e) => log.error('Failed to setup metrics server', e));
+}
 
 // Manual job execution via npm run jobs -- --execute <name>
 if (process.argv.length >= 4 && process.argv[2] === '--execute') {
